requirements.txt
python-telegram-bot==21.6
qrcode==8.2
Pillow==11.0.0
